requirements.txt
astroid==3.2.4
blinker==1.8.2
cached-property==1.5.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cramjam==2.8.3
curdleproofs==0.1.1
cytoolz==0.12.3
dill==0.3.8
eth-hash==0.7.0
eth-typing==3.5.2
eth-utils==2.3.1
eth2spec @ file:///home/nero/python/ethereum-explorer/minimal-ethereum-explorer/consensus-specs
Flask==3.0.3
Flask-Cors==5.0.0
hexbytes==1.2.1
idna==3.8
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
lru-dict==1.2.0
marko==1.0.2
MarkupSafe==2.1.5
mccabe==0.7.0
merlin-transcripts==0.1.1
milagro-bls-binding==1.9.0
mypy-extensions==1.0.0
platformdirs==4.2.2
py-ecc==6.0.0
py_arkworks_bls12381==0.3.4
pycryptodome==3.20.0
pylint==3.2.7
python-snappy==0.7.3
remerkleable==0.1.28
requests==2.32.3
rlp==3.0.0
ruamel.yaml==0.17.21
setuptools==74.1.1
sortedcontainers==2.4.0
tomlkit==0.13.2
toolz==0.12.1
trie==2.0.2
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
wheel==0.44.0
